requirements.txt
tornado==6.1
aiomysql==0.1.0
aiofiles==0.8.0
pyjwt==2.3.0
peewee==3.14.10
peewee-async==0.7.0
aioredis==1.3.1
Pillow==9.1.0
requests==2.27.1
aiocache==0.11.1
pycryptodome==3.14.1
paypalrestsdk==1.13.1
redisearch==2.1.1